requirements.txt
altgraph==0.17.2
asgiref==3.5.2
async-generator==1.10
atomicwrites==1.4.1
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==37.0.4
decorator==4.4.2
distlib==0.3.5
Django==4.1.4
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.8.0
future==0.18.2
h11==0.13.0
huepy==1.2.1
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
iniconfig==1.1.1
instabot==0.117.0
mock==4.0.3
moviepy==1.0.3
mysql-connector-python==8.0.31
numpy==1.23.1
outcome==1.2.0
packaging==21.3
pefile==2022.5.30
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
proglog==0.1.10
py==1.11.0
pycparser==2.21
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.28.1
requests-toolbelt==0.9.1
responses==0.21.0
schedule==1.1.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tabulate==0.8.10
tomli==2.0.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
urllib3==1.26.9
virtualenv==20.16.3
virtualenv-clone==0.5.7
wsproto==1.1.0
